Move BigQueryGCE to its own file (#3356)
* Move BigQueryGCE to its own file

* Add missing import
